Simplify PR
